% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{cfa_move_files}
\alias{cfa_move_files}
\title{Short cut to move all files in system.file directory to newly created
directory}
\usage{
cfa_move_files(fromDir, toDir)
}
\arguments{
\item{fromDir}{A \code{character} path where files are currently located}

\item{toDir}{A \code{character} path where you want the files to be}
}
\description{
Short cut to move all files in system.file directory to newly created
directory
}
