% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{cfa_sys_file}
\alias{cfa_sys_file}
\title{\code{system.file()} wrapper}
\usage{
cfa_sys_file(path)
}
\arguments{
\item{path}{A \code{character} path of target file relative to
\code{inst} (in development mode).}
}
\description{
Use to access files in \verb{inst/}.
}
\keyword{internal}
