% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_site_structure.R
\name{create_site_structure}
\alias{create_site_structure}
\title{Create Project Site Structure for Code for America Research Projects}
\usage{
create_site_structure(type, subfolder = ".")
}
\arguments{
\item{type}{The type of site to create, currently supporting only "cfa" as a valid type.
This allows for future expansion to support different project templates.}

\item{subfolder}{The subfolder in which to create the site structure. Defaults to the top-level directory.}
}
\value{
Invisibly returns \code{NULL}. The function is executed for its side effects
of creating files and directories, without a need for a return value.
}
\description{
This function automates the creation of a file and directory structure
specifically designed for hosting research projects. It facilitates the rapid
deployment of project content by organizing files and directories in a manner
compatible with static site generators like Quarto.
}
\details{
The function supports predefined types of sites, currently limited to "cfa"
projects. It performs parameter validation to ensure compatibility and prevent
errors during the structure setup.

Upon execution, the user is prompted to confirm the creation process to avoid
unintended directory changes. The function then proceeds to create a series
of directories and files:
\itemize{
\item Directories for R, site content (_site), logos, and fonts are created.
\item A custom CSS file (\code{styles.css}) is generated, pre-populated with basic
styling and custom fonts.
\item A set of Quarto markdown files (\code{.qmd}) are created for various sections
of the research project, such as executive summaries, analysis sections,
and data dictionaries. These files include default YAML headers and are
ready for further customization.
\item A \code{.gitignore} file to exclude certain data and temporary files from version
control.
\item Lastly, the function configures a \verb{_quarto.yml} file to define the site's
structure, navigation, and thematic elements for the site.
}

It's important to note that while the function provides a foundation for site
structure and initial file setup, further customization of content, styling,
and functionality is expected to be carried out by the research team
according to specific project needs and design guidelines.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
create_site_structure(type = "cfa")
\dontshow{\}) # examplesIf}
}
